PCSelectors, new actions, bugfixes
Added a new PCSelector dialogue item that can be used to allow the
player to choose a character to perform a action.

Added support for PCSelectors to the dialogue editor.

Added a few more core actions.

Made GiveItem and DisplayContainerInventory more flexible.

SkillChecks can now be rolled with more then one SkillCheckModifier
supplied.

Fixed a bug with skills not being written to XML properly.