AI refactoring, bug fixes.
Moved AI-related logic from GameCharacter to its own class, Brain.

Fixed a bug in PlayStateAnimationAction that would cause it to fail to
load the state.

Fixed a bug with Factions that were not getting reset on game unload.

Fixed a bug with murders not getting properly witnessed.

Fixed a bug with character circles not changing color when characters
become hostile.